Refactor layout editor into modular components
