+++ src/apps/layout/editor/utils.ts
@@ -1,0 +1,76 @@
+// src/apps/layout/editor/utils.ts
+import { LayoutEditorSnapshot, LayoutElement } from "./types";
+import { isContainerType } from "./definitions";
+
+export function clamp(value: number, min: number, max: number): number {
+    return Math.min(Math.max(value, min), max);
+}
+
+export function cloneLayoutElement(element: LayoutElement): LayoutElement {
+    return {
+        ...element,
+        attributes: [...element.attributes],
+        options: element.options ? [...element.options] : undefined,
+        layout: element.layout ? { ...element.layout } : undefined,
+        children: element.children ? [...element.children] : undefined,
+    };
+}
+
+function arraysAreEqual<T>(a: readonly T[] | undefined, b: readonly T[] | undefined): boolean {
+    const arrA = a ?? [];
+    const arrB = b ?? [];
+    if (arrA.length !== arrB.length) return false;
+    for (let i = 0; i < arrA.length; i++) {
+        if (arrA[i] !== arrB[i]) return false;
+    }
+    return true;
+}
+
+export function elementsAreEqual(a: LayoutElement, b: LayoutElement): boolean {
+    if (a === b) return true;
+    if (
+        a.id !== b.id ||
+        a.type !== b.type ||
+        a.x !== b.x ||
+        a.y !== b.y ||
+        a.width !== b.width ||
+        a.height !== b.height ||
+        a.label !== b.label ||
+        a.description !== b.description ||
+        a.placeholder !== b.placeholder ||
+        a.defaultValue !== b.defaultValue ||
+        a.parentId !== b.parentId
+    ) {
+        return false;
+    }
+    if (!arraysAreEqual(a.options, b.options)) return false;
+    if (!arraysAreEqual(a.attributes, b.attributes)) return false;
+    if (!arraysAreEqual(a.children, b.children)) return false;
+    if (!!a.layout !== !!b.layout) return false;
+    if (a.layout && b.layout) {
+        if (a.layout.gap !== b.layout.gap || a.layout.padding !== b.layout.padding || a.layout.align !== b.layout.align) {
+            return false;
+        }
+    }
+    return true;
+}
+
+export function snapshotsAreEqual(a: LayoutEditorSnapshot | undefined, b: LayoutEditorSnapshot | undefined): boolean {
+    if (!a || !b) return false;
+    if (
+        a.canvasWidth !== b.canvasWidth ||
+        a.canvasHeight !== b.canvasHeight ||
+        a.selectedElementId !== b.selectedElementId ||
+        a.elements.length !== b.elements.length
+    ) {
+        return false;
+    }
+    for (let i = 0; i < a.elements.length; i++) {
+        if (!elementsAreEqual(a.elements[i], b.elements[i])) return false;
+    }
+    return true;
+}
+
+export function isContainerElement(element: LayoutElement): element is LayoutElement & { children: string[] } {
+    return isContainerType(element.type) && Array.isArray(element.children);
+}
